uv_project/requirements.txt
langchain 
langchain-community
beautifulsoup4
langchain-text-splitters
sentence-transformers   #for embedding the data
langchain-huggingface
faiss-cpu